move component CSS styles and include as static styles
