+++ packages/calcite-components/src/components/block/block.scss
@@ -134,8 +134,6 @@
 
 @include includes.disabled();
 
-@include meta.load-css("../../styles/header");
-
 .header {
   @apply justify-start;
 }
