+++ packages/calcite-components/src/components/tip/tip.scss
@@ -32,7 +32,7 @@
   @apply m-0 border-none p-0;
 }
 
-@include meta.load-css("../../styles/header");
+@include meta.load-css("../../styles/component/header");
 
 .header {
   @apply mb-2;
