+++ packages/calcite-components/src/styles/component/header.scss
