+++ packages/calcite-components/src/styles/component/sortable.scss
